:art: Move ErrorPopup to dedicated folder
* Create and import styles file
