+++ client/src/components/error-popup/ErrorPopup.scss
@@ -1,0 +1,13 @@
+/**
+ * @ Author: Jbristhuille
+ * @ Create Time: 2022-12-24 15:00:00
+ * @ Description: Error popup styles
+ */
+
+/* SUMMARY
+*/
+
+.wcs-error-popup {
+    color: white;
+    background: var(--danger-color);
+}
